refactor: 処理, modelをファイルに切り出し(#12)
